requirements.txt
streamlit>=1.31.0
litellm>=1.40.0
jinja2>=3.1.0
playwright>=1.40.0
python-dotenv>=1.0.0
pyyaml>=6.0
Pillow>=10.0.0
